Added time server , firewall and remaining windows
