+++ api/python/provisioner/ui/confirm_dhcp_bmc.py
@@ -19,6 +19,6 @@
 from question_window import QuestionWindow
 
 
-class PrimaryWindow(QuestionWindow):
+class DHCPBmcWindow(QuestionWindow):
 
-    _question = "Is this the first node configured for this new cluster?"
+    _question = "Are you sure you want to set DHCP for BMC network?"
