+++ api/python/provisioner/ui/confirm_firewall.py
@@ -19,6 +19,6 @@
 from question_window import QuestionWindow
 
 
-class PrimaryWindow(QuestionWindow):
+class FirewallWindow(QuestionWindow):
 
-    _question = "Is this the first node configured for this new cluster?"
+    _question = "Are you sure you want to configure firewall?"
