+++ api/python/provisioner/ui/static_data.py
@@ -16,156 +16,12 @@
 # please email opensource@seagate.com or cortx-questions@seagate.com.
 #
 #
-import curses
-import config
-from window import Window
-from color_code import ColorCode
-from text_box import TextBox
-from success import SuccessWindow
+from form_window import FormWindow
 
 
-class StaticNetworkWindow(Window):
+class StaticNetworkWindow(FormWindow):
 
     data = {'Ip': '10.10.10.11',
             'Netmask': '1.255.255.251',
             'Gateway': '198.162.0.1'}
     component_type = 'Data Network'
-
-    def submit_button(self, x, y, menu, selected_rows):
-        buttons = {
-           "Submit": x + 15,
-           "Cancel": x + 5
-        }
-        selected_button = None
-
-        if selected_rows == len(menu):
-            selected_button = "Submit"
-        elif selected_rows == len(menu) + 1:
-            selected_button = "Cancel"
-
-        for button in buttons.keys():
-            if selected_button == button:
-                cod = ColorCode.get_color_pair(1)
-                self.on_attr(cod)
-                self._window.addstr(y + 1, buttons[button], button)
-                self.off_attr(cod)
-            else:
-                self._window.addstr(y + 1, buttons[button], button)
-
-    @staticmethod
-    def parse_input(kwargs):
-        if 'selected' not in kwargs:
-            selected_rows = 0
-        else:
-            selected_rows = kwargs['selected']
-
-        if 'edit' not in kwargs:
-            edit = False
-        else:
-            edit = kwargs['edit']
-        return selected_rows, edit
-
-    def create_window(self, **kwargs):
-        color_code = kwargs['color_code']
-        self._parent = kwargs['component']
-
-        selected_rows, edit = StaticNetworkWindow.parse_input(kwargs)
-        self._window.border()
-        col_code_attr = ColorCode.get_color_pair(color_code)
-        self.create_menu_head()
-
-        x = self.get_max_width() // 2
-        y = self.get_max_height() // 3 - 1
-        self.enable_keypad()
-
-        self.on_attr(col_code_attr)
-        self._window.addstr(y, 3, f"Please enter {self.component_type}"
-                            " static ips for this machine")
-        self.off_attr(col_code_attr)
-
-        values = list(self.data.keys())
-        count_m = len(values)
-        x = 6
-
-        for idx, row in enumerate(values):
-            y = self.get_max_height() // 3 - count_m // 2 + (idx + 1) * 2
-            if not (idx == selected_rows):
-                self._window.addstr(y, x, f"{row}:")
-                self._window.addstr(y, 14, f" {self.data[row]}")
-
-        y = self.get_max_height() // 3 - count_m // 2 + (count_m + 1) * 2
-
-        # Cancel and Submit button
-        self.submit_button(x, y, values, selected_rows)
-
-        if selected_rows < count_m:
-            y = (self.get_max_height() // 3 -
-                 count_m // 2 + (selected_rows + 1) * 2)
-
-            self.on_attr(col_code_attr)
-            self._window.addstr(y, x-3, ">> ")
-            self._window.addstr(y, x, f"{values[selected_rows]}:")
-
-            if edit:
-                user_content = TextBox(
-                    self._window,
-                    1,
-                    16,
-                    y,
-                    x + 10,
-                    self.get_max_height() // 4
-                ).create_textbox(
-                    color_code,
-                    self.data[values[selected_rows]]
-                )
-                self.data[values[selected_rows]] = user_content
-            else:
-                self._window.addstr(y, 14,
-                                    f" {self.data[values[selected_rows]]}")
-            self.off_attr(col_code_attr)
-
-    def process_input(self, color_code):
-        current_row = 0
-        values = list(self.data.keys())
-
-        while 1:
-            key = self._window.getch()
-            self._window.clear()
-
-            if key == curses.KEY_UP and current_row > 0:
-                current_row = current_row - 1
-
-            elif key == curses.KEY_DOWN and current_row < len(values):
-                current_row = current_row + 1
-
-            elif current_row == len(values) and key == curses.KEY_LEFT:
-                current_row = current_row + 1
-
-            elif current_row == len(values) + 1 and key == curses.KEY_RIGHT:
-                current_row = current_row - 1
-
-            elif key == curses.KEY_ENTER or key in (config.Key.EXIT_1.value,
-                                                    config.Key.EXIT_2.value):
-                if current_row < 3 and current_row >= 0:
-                    self._window.clear()
-                    self.create_window(
-                        color_code=color_code,
-                        selected=current_row,
-                        component=self._parent,
-                        edit=True
-                    )
-                    self._window.refresh()
-                elif current_row == len(values):
-                    win = SuccessWindow(self._window)
-                    win.create_window(
-                        color_code=2,
-                        data=f"Network data : {str(self.data)}"
-                    )
-                    break
-                else:
-                    break
-
-            self._window.clear()
-            self.create_window(color_code=color_code,
-                               selected=current_row, component=self._parent)
-            self._window.refresh()
