+++ api/python/provisioner/ui/static_mgmt.py
@@ -16,10 +16,10 @@
 # please email opensource@seagate.com or cortx-questions@seagate.com.
 #
 #
-from static_data import StaticNetworkWindow
+from form_window import FormWindow
 
 
-class StaticMGMTNetworkWindow(StaticNetworkWindow):
+class StaticMGMTNetworkWindow(FormWindow):
 
     data = {'Ip': '10.10.10.12',
             'Netmask': '11.255.255.252',
