+++ api/python/provisioner/ui/time_server.py
@@ -16,9 +16,11 @@
 # please email opensource@seagate.com or cortx-questions@seagate.com.
 #
 #
-from question_window import QuestionWindow
+from form_window import FormWindow
 
 
-class PrimaryWindow(QuestionWindow):
+class TimeServerWindow(FormWindow):
 
-    _question = "Is this the first node configured for this new cluster?"
+    data = {'Time Server': 'time.seagate.com',
+            'Time Zone': 'IST'}
+    component_type = 'Time server'
